requirements.txt
pandas==2.2.2
pyodbc==5.1.0
psycopg2==2.9.9
python-dotenv==1.0.1
loguru==0.7.2  
